Conclusão do cálculo "Long Method". Por terminar "Feature Envy".

+++ ES_2019_2020_Project/src/main/java/ES1_2019_METI_112/ES_2019_2020_Project/MethodDefinition.java
@@ -68,7 +68,7 @@ public class MethodDefinition {
 	}
 	
 	public boolean longMethodDefinition(String fileLOC, String fileCYCLO, 
-			int selectedLOC, int selectedCYCLO) {
+			double selectedLOC, double selectedCYCLO) {
 		boolean response;
 		if(Double.parseDouble(fileLOC) > selectedLOC && 
 				Double.parseDouble(fileCYCLO) > selectedCYCLO) {
